+++ NEWS.md
@@ -4,6 +4,10 @@
 
 - Columns automatically recorded at milestones can be eliminated from `get_output()` by setting new argument `tidy = TRUE`. 
 
+## Minor Updates
+
+- Some minor fixes for CRAN submission. 
+
 # TrialSimulator 1.6.0
 
 ## New Feature